refactor(app-server): move thread state types into module
